requirements.txt
jupyterlab==3.2.6
matplotlib==3.5.1
seaborn==0.11.2
numpy==1.22.0
pandas==1.3.5
scikit-learn==1.0.2
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0
Keras==2.8.0
tensorflow-hub==0.12.0
protobuf==3.20.1
parsenvy==3.0.2
